+++ NEWS.md
@@ -1,6 +1,7 @@
 # ggforce (development version)
 
 * Fixed a bug that would cause reordering of data in some geoms (#314)
+* The concaveman package is no longer a dependency for `geom_mark_hull()` (#308)
 
 # ggforce 0.4.1